backend/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
google-cloud-storage==1.42.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
newspaper3k==0.2.8
requests==2.32.3
PyJWT==2.9.0
werkzeug
lxml[html_clean]==4.9.3
flask-cors==3.0.10
tensorflow==2.12.0 